Rename package to vnet
