+++ lib/teleterm/vnet/service.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package service
+package vnet
 
 import (
 	"context"
